code/python/pyproject.toml
[project]
name = "hello-computer"
version = "0.1.0"
description = "Azure AI Speech demos for the Hello, Computer presentation"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech",
    "openai",
    "streamlit",
    "audio-recorder-streamlit",
    "python-dotenv",
]

[project.scripts]
hello-computer = "cli:main"
